fix: convert basicSalary from int to double for SalaryEntitlement

+++ mobile/lib/services/salary_entitlement_service.dart
@@ -112,7 +112,7 @@ class SalaryEntitlementService {
       employee: employee,
       hireDate: hireDate,
       totalMonthsWorked: totalMonthsWorked,
-      basicSalary: employee.basicSalary,
+      basicSalary: employee.basicSalary.toDouble(),  // ⭐ int to double
       totalEntitlement: totalEntitlement,
       totalWithdrawals: totalWithdrawals,
       totalDeductions: totalDeductions,
